[bin/pdfviewer] this (partially) fixes the bug described in [1]
One problem was that pdflatex fires an close_write event
directly before deleting the file. Therefore the existance
of the file is now checked before sending the refresh signal.
However this might not be enough as the file could be delete
right after the check.

The second problem was that if inotify watches the file
directly it isn't able to recieve any event after the
file was deleted, e.g. creation events. This is fixed by
watching the folder that contains the pdf and by filtering
out the right events.

[1] http://stackoverflow.com/questions/15451114/auto-reload-pdf-viewer-while-editing-latex-files